maxctrl: Use === and not == in comparison
With == you may get unwanted type coersion. E.g. (0 == '') evaluates
to true while (0 === '') evaluates to false. In this case the result
is that it is impossible to set a value to 0.

Incidentally, I do not understand what the (v == 'null') comparison
tests for. As far as I can tell it tests whether v is the string "null",
which seems nonsensical.