MXS-2729 Check buffer ownership only if rebalancing disabled
If some router or filter stores or has made a clone of a GWBUF
and rebalanching is done, then it may appear as if we would be
handling a GWBUF from the wrong worker. Currently it is not
possible to track those GWBUFs in order to update the owner,
so if rebalanching is enabled the owner will not be checked.