MXS-2729 Balance load of workers
The main worker will now at regular intervals wake up and check
the load of all workers. If the load of some worker is significantly
(as specific by the config parameter) higher/lower than the average
load, then a session will be moved from/to that worker.

Currently the period is hard-wired, but that will be made configurable.
Some stickiness will also be added so that the same session does not
end up being mover over and over again from one worker to the next.