Check `main` in newSession
The main target must be inspected in newSession to make sure that both
`main` and `targets` are still valid.

Ideally, the `main` parameter would depend on the `targets` parameter and
this would be caught during the configuration or any modifications. Due to
its special role, the `targets` parameter is processed after the service
is created, forcing the check to be made at session creation time.