Wait until all replies have returned
The report is generated when the last reply to the query has been received
but the result can be returned to the client before that. This means that
the last statement executed is not guaranteed to generate a report if the
main target is fast enough. To prevent this, the routing of the last chunk
of the response can be delayed until all replies have been received.

This change imposes an unavoidable delay in the case when one of the
targets is slower than the others. Whether this is an issue remains to be
seen but a less costly method of handling it would be to delay the closing
of the session up until the report is generated.