MXS-2691 Add session cache
A SessionCache instance is merely a thin wrapper around Cache
that in addition contains a token allowing the cache and storage
to distinguish between different users (i.e. sessions) of the
cache within the same thread.

This is needed so that we can arrange things so that the cost
for e.g. fetching something from a remote cache is payed by the
session in question and not by *all* sessions that happen to be
on the same thread.

This is but the first step; the cache API must be changed not only
so that the caller always must provide the token but also so that
the cache/storage can return a value indicating that the result
will be returned later to a callback.

The cache filter session itself must be modified so that if the
cache result will be returned to a callback, the session must
put itself and the client on hold until the result is delivered.