MXS-2505: Keep idle sessions alive
By changing the connection keepalive function to double as the idle
session callbacks, the connection keepalive also takes care of completely
idle sessions. Checking the idle time twice as often as necessary helps
make the keepalive more consistent. This could be done every time the
function is called but previously it has shown to incur a small
overhead. Whether this overhead is still present, and relevant, needs to
be measured.

The callback will be called more often than needed for sessions that are
completely idle but this overhead should not be significant as the
iteration over the backends is not always done.