MXS-2691 Prepare for async [put|del]_value
As far as caching a value, the code now works irrespective of
whether the cache is capable of storing a value immediately or
after some time.

In case the putting is done asynchronously, we do not wait for
that operation to finish before we return the result to the client.

The tricky thing here is that we need to cater for the case that
the session is closed before the callback is called. Hence 'this'
cannot be captured, but we need to capture the relevant variables
by value.