MXS-2505: Add generic connection keepalive
The connection keepalive is now done in a way that is not dependent on a
specific router implementation. One of the downsides of this is that it
now relies on request tracking which means an additional overhead for
those routers that don't need it (readconnroute). The move to periodic
checks also causes a few extra function calls for each session roughly
once every second but, since this was already done for the idle connection
timeouts, it should not incur a noticeable overhead. If it does, the cost
can be further amortized by spreading out the individual timeout checks.

With the move to a generic mechanism, the need for the timeout style
callbacks is removed. This makes code clearer and it keeps all the higher
level functionality nicely contained inside the Session class. The only
external part is the new ping method that backend connections must
implement.

The current pinging mechanism uses the old ignorable ping mechanism which
is not really built for this specific purpose. A more refined mechanism
that understands when to discard results would make the code clearer.
This also means that the request tracking must be enabled in order to know
when a connection is logically idle. For example, a SELECT that takes two
hours to complete must not cause a keepalive ping to be generated as the
connection is not logically idle. In addition to this, it would also be
both unnecessary as well as hard to handle with the current
implementation.

The check for idle connections is done twice per configured interval. This
gives a worst-case latency of about 1.5 times the configured
interval. Further reducing the check interval would improve the accuracy
but it would also increase the overhead that the checking imposes.