MXS-2691 Change Filter<T,S> api function names
The functions of Filter<T,S> that implement the API defined by
MXS_FILTER now has an api-prefix.

The reason is that the class derived from Filter - e.g.
class XFilter : public Filter<XFilter, XFilterSession> - is
expected to implement a functions, e.g. 'newSession()' whose prototype
is different from the Filter<T,S>::newSession that it hides.

This was not apparent earlier because in the object defining the API
the template functions, e.g.

    &Filter<FilterType, FilterSessionType>::newSession

was used. That is, it was not possible to compile time override that
implementation of the API function. That will now be necessary in the
Cache, which means that the above must be like

    &FilterType::newSession

which does not work due to the hiding.

So, to deal with this, the API-functions are now prefixed with 'api'.