MXS-2505: Make keepalive interval private to Session
The session now calculates the need for a keepalive ping by calling an API
method that returns the number of seconds the connection has been idle
for. This removes the need to check the connection keepalive values both
in the session and in the backend connection. In addition to this, it
greatly improves the responsiveness of keepalive pings.