MXS-2691 Allow creation of std::weak_ptr to CacheFilterSession
When fetching cached data asynchronously, the following chaing of
events is possible:

1) SELECT arrives.
2) An asynchronous call to the cache is initiated.
3) The client session is put on hold.
4) The client closes the connection, thus forcing the session to be
   closed.
5) The cache callback arrives.

At point 5) we need to be able find out whether the CacheFilterSession
is still alive.

CacheFilterSession now contains a shared_ptr member referring to itself.
At session free time we arrange it so that a copy of the member is made,
the member is reset and the copy is allowed to go out of scope. That
allows the instance to be deleted, without it leading to a double delete.

Now, from the shared_ptr we can create a weak_ptr that we store along
with the asynchronous call (in a lambda). When the lambda is called
we lock the weak_ptr, which gives us a shared_ptr, and then from that
shared_ptr we can find out whether the CacheFilterSession is still alive
or not and act accordingly.