MXS-2691 Prepare for asynchronous Cache::get_value
The callback to the asynchronous get_value() has a weak pointer
to the CacheFilterSession. That allows us at callback time to
figure out whether the session still exists.

The code for deciding whether to return the response immediately
or to continue routing (i.e. deliver the packet downwards) is
the same, however, the way a found response is delivered upwards
has to be different.

In the synchronous case, we call set_response() that calls
session_set_response(), which will cause the response to be
delivered when the routeQuery() processing returns.

In the asynchronous case, we call the clientReply() of the up
component directly.