MXS-2691 Prepare for shared storages
If invalidation is required, the storage is shared and the storage
does not natively support invalidation we must fail.

Also, configuration settings for capping the size and the maximum
number of items in the cache are ignored if the storage is shared.
That's something that has to be specified when the shared storage
is setup.