MXS-2708 Provide token implementation for storage_memcached
Now each session will have an actual token of its one. The token
contains a handle to memcached and thus it means that the storage
imlementation now is multi-thread capable.

Memcached is still used in a synchronous fashion so session handled
by the same worker will still have to wait for each other (when a
[put|get|del]_value() is executed), but sessions handled by other
workers will not.