MXS-2708 Implement asynchronous memcached storage
Contrary to initial belief, it is not possible to use libmemcached
in an asynchronous fashion. The socket can be made non-blocking,
but in that case libmemcache internally loops using poll() so to
an external user it still appears synchronous. You can cheat a little
so that the first read is done only when there is something to read,
but that buys little if there is a lot of data to read and nothing at
all when putting or deleting values.

So, all interaction with memcached is now done in a thread that when
needed is created for the purpose. This will be changed:
- first so that each each session has a dedicated thread to be used for
  all communication (there can only ever be one put/get/del operation
  going on)
- then so that a generic thread-pool is added. As not every session
  will all the time have a memecached operation in progress it will
  significantly reduce the number of threads created.