Add TTL for sessions
By introducing a TTL mechanism for sessions, we guarantee that all
affected sessions will be closed. If they are close before the TTL is
reached, the session terminates normally. If the session lifetime exceeds
the TTL, the session is forcibly closed.

This mechanism can be used as a building block to guarantee that sessions
started before a configuration change terminate within an expected
period. This can also be used to perform a controlled shutdown where no
new sessions are accepted but old sessions are allow to close gracefully.