MXS-2827 Allow repeated user account refreshes at startup
Throttling only kicks in after users have been loaded a few times,
or the loading has failed several times in a row. This should
help with tests that create users at the beginning.

Even without throttling there is a small delay (500ms) between
user refreshes, and during this time authentication may fail.
The delay can be shortened if it's causing tests to fail.