MXS-2725: Add max_connections to diagnostic output
Moving the statistics object creation inside mxs::Target::Stats allows
both services and servers to share the same code.

The service resources were missing the statistics member that was present
in server resources. To keep the layout of the resources similar, the
services now have a statistics member in the attributes object that
duplicates some of the data present in other members of the attributes
object. This retains backwards compatibility with older versions.

Removed duplicated information from readconnroute diagnostic output.

Added the max connection count to maxctrl and changed the REST API calls
to use the new statistics member of the attributes object for both
services and servers.