MXS-2819 Use MULTI when updating redis
When storing stuff, the updating of the set with the keys and the
storing of the value itself, is done within one transactions.

When invalidating, the clearing of the set and the deletion of the
values is done within one transaction (the fetching of the keys is
done separately).

This is not quite sufficient yet to make invalidation work in all
circumstances, as the set may be modified between fetching and deletion
of keys.