MXS-2819 Use redis WATCH
Now the WATCH functionality of redis is used for ensuring that
a concurrent invalidation and cache insertion are handled in a
fashion that ensures that the cache reflects what actually is
in the server.

This is an optimistic locking scheme, which in principle means
that one particular attempt to store a value to the cache
(put_value()) or to invalidate values (invalidate()) may never
get a chance to do so.

Testing will need to be performed in order to fine out how likely
it actually is.