Move all replication processing into Rpl
The Rpl class can now be used as a generic class that handles MariaDB
replication and converts it into row events with plain C++ types. To use
the class, one has to implement the functionality to read individual
replication events as well as provide an implementation for the
RowEventHandler interface.