Add basic mariadb-to-kafka replication
The kafkacdc router uses the replicator to stream the replicated events as
JSON directly to Kafka. This is an improvement over the current method of
using the avrorouter and a script to pipe the scripts in to Kafka by
removing the need for on-disk data.

Due to the simple nature of the implementation, a crash of MaxScale can
cause missing or duplicated events.