MXS-2774 Allow specifying whether param can be modified at rt
Now when a parameter is declared, it is possible to specify whether
it can be modified at runtime or not.

That will, on the one hand, be used in config_runtime.cc when deciding
whether to allow the modification of a parameter, and, on the other
hand, be used when the parameter value is accessed for deciding whether
the atomicity explicitly need to be ensured.