MXS-2774 Expose configuration via the REST-API
The core parameters are now exposed as if the core would
be a module in itself. They are also explicitly accessible
via /v1/maxscale/modules/core. Currently there is a
discrepancy between how the core parameters and the parameters
of other modules appear, and that is due to the fact that the
core parameters are now exposed via the new configuration
mechanism while the others are exposed via the old.