Fix auth errors during startup
The user data updating would not update users right after startup. To
allow this, the last update should be set in the past.

Starting the monitors before services and waiting for one full monitor
interval before the services are started prevents sessions from failing
due to the server being in an invalid state.

Also fixed the npm test script to kill the process if it's still running.