MXS-2793: Fix processing of multi-table queries
By storing all locations where a particular table or database exists the
query processing can pick the best target at the time the query is made
instead of when the database mapping is done.

This commit removes the only-one-location rule for databases which breaks
the ignore databases as well as the preferred server mechanisms. The
former needs to be fixed in a separate commit and the latter is made
redundant by this change and thus it only needs to be ignored.