MXS-1720: Store GTID in servers
The GTID sequence numbers can now be stored in the server. Due to the
simple nature of the implementation, only 32 different GTID domains are
supported. Although it is unlikely that this value is exceeded in normal
use, the algorithm could still be adjusted to forcibly eject one of the
domains when a free slot is not found. This idea still needs to be
evaluated.