Fix connection limit off-by-one-error
The connection counts are incremented only when the session has been
successfully started. This means that the limit must not have been reached
when the newest connection is created.

There is an inherent race condition in counting only fully authenticated
connections. It is possible that the number of created connections exceeds
the number of max_connections if multiple connections are created at the
same time.