MXS-2145 Add connection_init_sql_file-setting
During service creation, the file is read into memory. When creating
backend connections, the contents of the file are sent as sql queries
to the backends after authentication is complete. The format of the file
is very simple, every line is sent as a single sql-statement. Each statement
should return an OK-packet.