MXS-2531 Add manual lock release command
This can be used to force a monitor to release any of its locks so that
another monitor can claim them. The releasing monitor will not try to
reacquire locks for 1 minute, so the command should only be used when
it's know that another monitor is observing the lock status.