Simplify target selection
If no target is found (i.e. a null pointer is returned) then it can always
be considered a failure. This again moves the logic of master assignment
away from the target selection code which in turn simplifies the latter
part.

Also fixed the transaction migration code not using the correct pointer in
the start_trx_migration call.