MXS-2922 Add 'admin_host' parameter to csmon
MaxScale will have to orchestrate (and provide means for administering)
Columnstore and for that there will be a REST-API. Currently it's
somewhat unclear exactly who will provide that REST-API; every Columnstore
UM or some independent component. Starting the Columnstore cluster surely
cannot be provided by the UMs and neither can failover (there's supposed
to be a REST-API call for that).

For the time being it is assumed there will be an administrative host,
the exposed REST-API of which will be used for administering and
orchestrating Columnstore. The Columnstore REST-API is work in progress
but there is a dummy POC implementation that responds to the current
calls.

But note that this is all very tentative and not necessarily as it
eventually will be. However, this will not be completely wrong.