Don't switch to basic auth after an invalid token
The WWW-Authenticate header returned by an invalid token triggers a modal
to pop up whenever token authentication is used. This is usually not the
desired behavior and a plain HTTP 401 without a recommended authentication
scheme in these situations is better.

Also fixed the logging of token validation errors not respecing
admin_log_auth_failures.