Pass null values forward
Previously the values were ignored and only updated values were passed to
the callbacks. Passing null values should be done if the JSON is used to
update the object as it allows us to know when a parameter has been
removed without having to resort to passing empty strings.

Adjusted the unit tests accordingly and removed their dependency on the
shared object that was mutated by each test step.