MXS-2922 Implement async commands separately
After a few attempts of doing it in a more generic fashion (a single
'async' command taking the actual command and a varying number of
parameters) a decision was made to follow the old wisdome "when in
doubt, use brute force", that is, use a simple and guaranteed to
work approach.

Currently at most one command - synchronous (obviously) or
asynchronous - can be in process at any one time. The result of an
asynchronous command must be fetched, before another command can be
issued.