MXS-2922 Add skeleton implementation of add-node
The current algorithm for adding a node is:

* cluster add node { IP | DNS }
  - Sends GET /node/ping to the new node.
    Fail the command  if the previous call failed.
  - Sends GET /node/config to { all | only one } of the listed nodes.
    Uses the config/-s to produce new versions of the configs.
  -Sends PUT /node/config to the old nodes and to the new node.
  The previous action forces the config reload for all services.

However, in order to do this properly MaxScale must be able to figure
out whether a server is part of the cluster only as far as MaxScale's
configuration goes or if it is part of the cluster both as far as
MaxScale's configuration *and* Columnstore's configuration goes. In
the former case the adding should be attempted, in the latter it should
cause an error.

Since this is the case, it seems that the CS monitor should be like the
Clustrix one; that is, the servers are only used as bootstrapping servers
using whih the monitor can connect to the cluster and then the actual
servers are created dynamically from the Columnstore configuration. That
would effectively remove the possibility that MaxScale and Columnstore
have a different view of what consititutes the cluster.