Remove post_configure from smartrouter
Since the check was previously done on the configuration parameters, the
configuration could be validated on startup. Now that the parameter types
are not passed to the post_configure due to there being two types of them,
an alternative is needed.

Most of the behavior of the smartrouter will remain the same with the
exception of the router not refusing to start if the master target is not
found. The sessions will still refuse to start which means the runtime
functionality is still the same.

This fixes the sts_smartrouter test.