MXS-2925: Fix yargs compatibility problems
Due to https://github.com/yargs/yargs/issues/1323 the arguments to
commands must not have the same name as any global option.

Removed custom errors in favor of auto-generated ones. For example `help
list` no longer works and must be `list help` or `list --help` instead.

The custom options given by tests are now injected before the command
instead of after it. This seems to be a required change with the new Yargs
version.

Fixed the test stop script to exit immediately if it sees that the
maxscale process is gone. This prevents stray SIGABRTs from causing false
positives.

Removed double spaces in cluster test that translated into an argument
that was an empty string.