MXS-2948: Make added/deleted diff more compact
Only the type and name of the deleted and added objects is now
returned. Given that the details of these particular objects are not the
main interest, this is enough to let the user know one MaxScale has a
different set of objects than the other.

Also fixed cluster diff unit tests to understand the new format.