MXS-2948: Fix the --rdns option and test it
Boolean types should use 'boolean' instead of 'bool'. Older versions of
yargs injected unknown options into the argv which explains why it works
in the old version. Newer versions of yargs treat unknown options as
errors so the type declaration must also be fixed.

Added a test case for the --rdns option that simply checks the option is
parsed properly. Proper testing for this would be harder but at least the
code is now covered by a test case.