MXS-2954: Use correct object for PATCH
The destination object instead of the source object was used. This caused
the PATCH to be a no-op.

Due to how the REST API behaves when it received parameters, the set of
synced parameters must be filtered. The REST API should be more lenient
towards alterations to parameters that don't actually change the value.