MXS-2924: Add listener endpoints
The /listeners endpoint returns listeners from all services and the
/listeners/:listener returns a specific listener. The listeners resource
now always has the service relationships regardless of whether it is a
separate listener resource or a sub-resource of a service.

The services also have a listeners relationship in addition to the
sub-resource. The sub-resource should be kept around for backwards
compatibility but all new code should use the new listeners endpoint.

Extended and fixed the test suite to validate that the new resource
endpoints conform to the JSON API specification.