MXS-2922 CsMonitorServer fetches but does not store status
Storing the status in CsMonitorServer is only likely to lead
to a difference in opinion of what the actually status is, so
now CsMonitorServer just provides a way for fetching it.
Further, all output related to the status is now handled by
CsMonitor. The config will go the same way.