Add workaround to connector-c rotate event problem
The connector doesn't correctly handle the name of the binlog in the
rotate event. In addition to this, we can't reliably extract it even if
was processed correctly as the checksum knowledge is not known by the
connector.

To work around the first problem, we process the raw event.

To work around the second problem, we could skip the assignment of the
@master_binlog_checksum server side variable but this would prevent
replication if binlog checksums are enabled on the master.

An uncomfortably effective hack to this problem is to calculate the crc32
of the event assuming that the last four bytes contain the crc32. If
binlog checksums are enabled, the calculated checksum will match the
stored one. If the calculated checksum doesn't match the "stored" crc32,
either we've read a corrupt binlog event or binlog checksums are not
enabled. If we ignore corrupt events, we can deduce whether checksums are
enabled using this method.