Integrate pinloki into MaxScale
Added the router boilerplate code required to build and link it against
the pinloki sources. The router now simply answers with an OK packet to
all queries.

Changed to mxb::Worker from the custom one for the Reader. The writer
still needs to be in its own thread as it doesn't have incoming events.