Store Config and Inventory in Pinloki
The use of singletons prevented multiple pinloki routers from being
created at the same time.

Pass a reference to an Inventory around and expose the configuration used
to create it. This allows only one object to be given to most constructors.

Moved the thread creation inside the Writer class. Changed Reader to use
the worker given to it as an argument instead of creating a new worker for
each reader.