Add START/STOP SLAVE functionality
START SLAVE now starts a new Writer if the configuration is correct and
STOP SLAVE stops it. CHANGE MASTER also persists the configuration to
disk.

The Writer is mostly functional with the exception being that stop events
followed by an artifical rotate event do not cause the binlog to be
rotated.