Add simple flow control
In order to prevent excessive memory usage, the reading of the binlogs
must be stopped when the client DCB's network buffer is full. The act of
stopping is not very complex but the act of continuing is a more involved
process.

The current implementation uses the delayed call mechanism to periodically
check whether the reading can be continued. A more advanced mechanism
would use a callback that would be called when the EPOLLOUT event drains
the writeq of the client DCB. The DCB callback mechanism could be used for
this but as it hasn't been modernized, it is not the best option.