MXS-2922 Update configuration on first node of cluster
On a single-node Columnstore cluster the value of IPAddr
will be 127.0.0.1. In order to turn it into the first node
in a multi-node cluster, we need to change 127.0.0.1 to an
actual IP.

Currently there's a lot of logging going on when things
succeed. Once things have been verified to work, excessive
logging will be pruned.