Add syntax for MASTER_GTID_WAIT
This will make synchronization of pinloki to a certain GTID position much
easier. This will mainly benefit testing but it can also be used to verify
that a transaction has been processed by pinloki.

The MASTER_GTID_WAIT rule isn't a part of the SELECT rule as that would
make the processing much more complex. This means that only the exact
query of SELECT MASTER_GTID_WAIT will match the MASTER_GTID_WAIT rule. If
the function is used together with another function, the literal string
will be returned. Whether this limitatino is a problem needs to be
estimated.