Prevent error message spam
If an exception, for example a network error, is repeatedly thrown, wait
for a short while before retrying. This will prevent flooding of the log
but it is not an optimal solution.

Whether the thread waits or not should depend on the type of the
error. Broken network connections and timeouts should trigger a
reconnection immediately, temporary network errors should be retried soon
and permanent failures should cause the replication to stop.

Also added more detail into the error messages to help convey why or from
where it originated and fixed the includes.