Fix current write position tracking
The write position was not correctly tracked when an existing file was
opened. Artificial events would overwrite the current write position even
if they were before m_sync_pos. Fixing this removes the need for the
change done to the file writer in commit
3ce588409e2dec52c8d794255780af5738ef32fb.

Removed the info level messages for end-of-file errors and added info
level messages when the file synchronization is started and stopped.