Add pinloki master change test case
The test checks that back and forth changes in masters work with pinloki
as long as the new master is always ahead of the old one. The test does
not currently pass as the STOP SLAVE followed by FLUSH LOGS will cause the
last binlog to not have the final STOP_EVENT or ROTATE_EVENT that closes
the file.