MXS-2922 Provide functionality for updating dbroots
Db roots can be added to a node by an administator and they
are subsequently reported by GET /node/status. However, the
configuration is not updated automatically.

So, the Columnstore monitor must be able to figure out whether
the currect configuration reflects the reality and if not, update
the config and update the cluster.