Wait for the first artificial GTID LIST before writing to binlogs.
Also in this commit, decide whether to processes a rotate event based
on event flags. This has been tested with pinloki restarts, slave server
restarts, STOP SLAVE and START SLAVE. There are a few special cases
that need to be added, which will come in later commits (master restart
and log flushes cause the master to send only artificial rotates in
some cases, which are currently ignored).