Fix TLS usage in Connection::connect
Renamed the method to keep it consistent with the rest of the code
base. Fixed the connection error generation which used a null MYSQL
handle.

Added the missing MYSQL_OPT_SSL_ENFORCE option. The code now also checks
that the connection is indeed encrypted if TLS was requested.

Fixed the test case to work with replication that requires TLS. In order
to do this, the setup method must be overridden.