Fix Replicator and test_kafkacdc
The heartbeat events that were recently added did not trigger
shutdown. This caused the test to generate a core dump when systemd killed
MaxScale.

After fixing this, the test would sporadically fail due to the consumer
being started after the events were generated. To guarantee that the
events are received, the consumer must be started before the events are
generated.