Prevent token auth and GUI without HTTPS
To prevent tokens from leaking, the auth endpoint must only be enabled
when HTTPS is enabled as well. This goes for the GUI as well since it uses
token authentication.

The error page for missing HTTPS when GUI is enabled should help people
enable TLS for the REST API or to disable the warning.