MXS-3067: Load all modules on startup
All libraries found in the MaxScale library directory that implement an
up-to-date version of the expected APIs are loaded.

This fixes the problem where the parameter information exported by the
REST API was only available after a module was already loaded. This means
that in order to know how to feed the parameters to the REST API, one
failed attempt to load a module was required.