Fix `admin_secure_gui`
The warning about the GUI not being served without HTTPS was still logged
and the /auth endpoint would refuse connections without HTTPS. Tied these
two features into the `admin_secure_gui` parameter as it's mainly related
to that.

Clarified the documentation to be clear that the `admin_secure_gui`
parameter affects the /auth endpoint.