2.3 Add script for setting change date
This script is capable of setting the change date and not just
updating a specific one.

It looks for " Change Date: .*" in the first 20 lines of file and
replaces whatever follows "Change Date: " with the specified change
date.

If provided with a directory, it will walk the directory hierarchy
and recursively set the change date in all files.