Revert v-main to v-content as vuetify@2.3.4 requires manual migration
In package.json, vuetify version is "^2.2.34", the caret means to install minor
versions update but instead of that, in vuetify, it installs 2.3.4 version with
requires manual migration and deprecated warning pops up.
Hence, changing caret to tidle to install new version of vuetify with bug fixes only.
Also bumping other packages version to be consistent with package.lock.json.