Improve navigating back function in details page
If unauthenticated or authenticated user accesses maxgui with
direct link to details page, previous route name will be login
or null, so when clicking the go back icon, it should not
go back to login page but dashboard page.

This commit adds a global prevRoute state before routing.
details-page-title uses this state to handle route changes.