MXS-3099 The default value for a server param must be NULL
The default value for an optional server parameter must
be NULL and not an empty string. If the latter, the dynamic
linking of a server to a monitor will fail.

Case in point, the Columnstore monitor has an optional server
parameter "primary". Without this change, a server cannot be
added dynamically to it.