Refactor genDataSetSchema action
Dataset schema action should commit mutation with
payload datasets as an array instead of an object.
This also sets line-chart legend of each dataset to have
background color the same as border-color. The
background color has 0.1 opacity, this makes it difficult
to see.

This reduces complexity for genDataSetSchema action and make
it easier for drying up this action in the next commit.