Fix defaultItems in select-dropdown
When creating a resource, if defaultRelationshipItems is defined,
resource failed to be created. This is because the component didn't
emit any events if defaultItems are defined.
In addition, server-form-input and service-form-input have two
resource-relationships component but using the same defaultItems.
This leads to different relationship types have the same values.

onChange method in the watcher of defaultItems needs to be called
to emit component's events.
This also separates defaultItems data in server-form-input and
service-form-input.