MXS-3083: Allow empty strings
The parameter validation for the REST API now accept empty strings which
allows certain default values to be correctly processed. This happened
with string parameters that had an empty value as the default. When these
parameters were automatically filled by the configuration code, it would
result in an invalid configuration.