MXS-3124: Fix COM_STMT_EXECUTE metadata injection
The metadata was always injected even if the COM_STMT_EXECUTE had no
parameters. In addition, the code caused a non-contiguous buffer to be
generated which would trigger a debug assertion.

If the router receives a COM_STMT_EXECUTE with parameters that does't have
the send types flag set, a previous COM_STMT_EXECUTE with it enabled must
have been sent. Previously the code created an empty value regardless of
whether the value was seen before. If this ever happens, a warning should
be logged as it indicates either a broken connector or a logical error in
readwritesplit.