MXS-2754: Replace configureInstance with getConfiguration
By returning a mxs::config::Configuration the MaxScale core can perform
all of the configuration validation as well as the actual act of updating
the configuration. Any module specific actions that need to be done can be
done in the post_configure method of the class. This is a better solution
than the old configureInstance entry point that requires more work from
the module.

This change also simplifies the whole configuration process by making the
initial configuration and any subsequent runtime reconfigurations behave
in the same way.