MXS-2754: Replace use of Specification::populate
The modules that use the new configuration mechanism now expose the
Specification in the MXS_MODULE. This allows the populate method to be
removed.

The stored pointer to the Specification should also be const.

Changed the cache filter default values to work without the quote removal
that was done by the old configuration code.